requirements.txt
# CORE ENGINE
torch --index-url https://download.pytorch.org/whl/cpu
transformers
peft
datasets
accelerate
trl

# DATA & MATH
pandas
numpy
scikit-learn
tqdm
faiss-cpu
sentence-transformers

# API BACKEND
fastapi
uvicorn[standard]
python-multipart
pydantic
